Move shared npm dynamic validation logic to NpmAuthorTokenHelper
